requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-auth
pydantic
python-dotenv
cloud-sql-python-connector[pg8000]
google-cloud-secret-manager
